Natural-Language-to-SQL ROC project/requirements.txt
langchain_community>=0.0.30
langchain-openai>=0.1.7
sqlalchemy>=2.0.0
pyodbc>=4.0.39
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.29.0

